Rename config to spark_config
