+++ immunopepper/mode_crosscohort.py
@@ -10,9 +10,9 @@ import shutil
 import sys
 from typing import Iterable
 
-from .config import create_spark_session
-from .config import create_spark_session_from_config
-from .config import default_spark_config
+from .spark_config import create_spark_session
+from .spark_config import create_spark_session_from_config
+from .spark_config import default_spark_config
 
 def organize_inputs(input_dir, backgr_suffix, kmer, samples, mutation_modes, compres_suffix, skip_reorganize):
     to_combine = []
